requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-emit-index-url requirements.in
#
alabaster==0.7.12
    # via sphinx
appdirs==1.4.4
    # via black
async-generator==1.10
    # via nbclient
attrs==20.3.0
    # via
    #   jsonschema
    #   pytest
babel==2.9.0
    # via sphinx
beautifulsoup4==4.9.3
    # via c4.spanner
black==20.8b1
    # via -r requirements.in
bleach==3.2.2
    # via nbconvert
boto3==1.14.44
    # via
    #   c4.spanner
    #   watchtower
botocore==1.17.63
    # via
    #   boto3
    #   c4.spanner
    #   s3transfer
bump2version==1.0.1
    # via -r requirements.in
c4.spanner==0.0.21
    # via -r requirements.in
certifi==2020.12.5
    # via requests
chardet==4.0.0
    # via requests
click==7.1.2
    # via
    #   -r requirements.in
    #   black
    #   c4.spanner
    #   pip-tools
coverage==5.3.1
    # via
    #   -r requirements.in
    #   pytest-cov
defusedxml==0.6.0
    # via nbconvert
docutils==0.15.2
    # via
    #   botocore
    #   sphinx
entrypoints==0.3
    # via nbconvert
fastparquet==0.5.0
    # via c4.spanner
flake8==3.8.4
    # via -r requirements.in
future==0.18.2
    # via c4.spanner
idna==2.10
    # via requests
imagesize==1.2.0
    # via sphinx
iniconfig==1.1.1
    # via pytest
ipython-genutils==0.2.0
    # via
    #   nbformat
    #   traitlets
jinja2==2.11.2
    # via
    #   nbconvert
    #   sphinx
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
jsonschema==3.2.0
    # via nbformat
jupyter-client==6.1.11
    # via nbclient
jupyter-core==4.7.0
    # via
    #   jupyter-client
    #   nbconvert
    #   nbformat
jupyterlab-pygments==0.1.2
    # via nbconvert
llvmlite==0.35.0
    # via numba
loguru==0.5.3
    # via -r requirements.in
markupsafe==1.1.1
    # via jinja2
mccabe==0.6.1
    # via flake8
mistune==0.8.4
    # via nbconvert
mypy-extensions==0.4.3
    # via black
nbclient==0.5.1
    # via nbconvert
nbconvert==6.0.7
    # via c4.spanner
nbformat==5.1.2
    # via
    #   c4.spanner
    #   nbclient
    #   nbconvert
nest-asyncio==1.4.3
    # via nbclient
numba==0.52.0
    # via fastparquet
numpy==1.19.5
    # via
    #   -r requirements.in
    #   c4.spanner
    #   fastparquet
    #   numba
    #   pandas
    #   pyarrow
    #   scipy
packaging==20.8
    # via
    #   bleach
    #   pytest
    #   sphinx
pandas==1.2.1
    # via
    #   -r requirements.in
    #   c4.spanner
    #   fastparquet
pandocfilters==1.4.3
    # via nbconvert
pathspec==0.8.1
    # via black
pip-tools==5.5.0
    # via -r requirements.in
pluggy==0.13.1
    # via pytest
psycopg2-binary==2.8.6
    # via c4.spanner
py==1.10.0
    # via pytest
pyarrow==2.0.0
    # via c4.spanner
pycodestyle==2.6.0
    # via flake8
pyflakes==2.2.0
    # via flake8
pygments==2.7.4
    # via
    #   jupyterlab-pygments
    #   nbconvert
    #   sphinx
pyparsing==2.4.7
    # via packaging
pyrsistent==0.17.3
    # via jsonschema
pytest-cov==2.11.1
    # via -r requirements.in
pytest==6.2.1
    # via
    #   c4.spanner
    #   pytest-cov
python-dateutil==2.8.1
    # via
    #   botocore
    #   jupyter-client
    #   pandas
python-dotenv==0.15.0
    # via -r requirements.in
pytz==2020.5
    # via
    #   babel
    #   pandas
pyzmq==21.0.1
    # via jupyter-client
regex==2020.11.13
    # via black
requests==2.25.1
    # via
    #   c4.spanner
    #   sphinx
s3transfer==0.3.4
    # via boto3
scipy==1.6.0
    # via c4.spanner
six==1.15.0
    # via
    #   bleach
    #   jsonschema
    #   python-dateutil
    #   thrift
snowballstemmer==2.1.0
    # via sphinx
soupsieve==2.1
    # via beautifulsoup4
sphinx==3.4.3
    # via -r requirements.in
sphinxcontrib-applehelp==1.0.2
    # via sphinx
sphinxcontrib-devhelp==1.0.2
    # via sphinx
sphinxcontrib-htmlhelp==1.0.3
    # via sphinx
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==1.0.3
    # via sphinx
sphinxcontrib-serializinghtml==1.1.4
    # via sphinx
sqlalchemy==1.3.22
    # via c4.spanner
testfixtures==6.17.1
    # via c4.spanner
testpath==0.4.4
    # via nbconvert
thrift==0.13.0
    # via fastparquet
toml==0.10.2
    # via
    #   -r requirements.in
    #   black
    #   pytest
tornado==6.1
    # via jupyter-client
traitlets==5.0.5
    # via
    #   jupyter-client
    #   jupyter-core
    #   nbclient
    #   nbconvert
    #   nbformat
typed-ast==1.4.2
    # via black
typing-extensions==3.7.4.3
    # via black
urllib3==1.25.11
    # via
    #   botocore
    #   requests
watchtower==1.0.6
    # via -r requirements.in
webencodings==0.5.1
    # via bleach

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
